draft substitutions
