+++ lib/substitutions/include/substitutions/graph_pattern.h
@@ -1,32 +1,24 @@
-#ifndef _FLEXFLOW_SUBSTITUTIONS_GRAPH_PATTERN_H
-#define _FLEXFLOW_SUBSTITUTIONS_GRAPH_PATTERN_H
+#ifndef _FLEXFLOW_SUBSTITUTIONS_SUBSTITUTIONS_H
+#define _FLEXFLOW_SUBSTITUTIONS_SUBSTITUTIONS_H
 
-#include "utils/graph.h"
+#include "graph_pattern_match.h"
 
 namespace FlexFlow {
 
-struct DiGraphPatternMatch {
-  bidict<Node, Node> nodeAssignment;
-  req<bidict<OpenMultiDiEdge, MultiDiEdge>> edgeAssignment;
+struct GraphPattern
+    : public strong_typedef<
+          GraphPattern,
+          LabelledOpenMultiDiGraph<OperatorPattern, ParallelTensorPattern>> {
+  using strong_typedef::strong_typedef;
 };
-
-FF_VISITABLE_STRUCT(DiGraphPatternMatch, nodeAssignment, edgeAssignment);
-
-struct MatchSplit {
-  DiGraphPatternMatch prefix_submatch;
-  req<DiGraphPatternMatch> postfix_submatch;
-};
-
-FF_VISITABLE_STRUCT(MatchSplit, prefix_submatch, postfix_submatch);
 
 GraphSplit split_pattern(OpenMultiDiGraphView const &pattern);
 
-bool pattern_matches(OpenMultiDiGraphView const &,
-                     MultiDiGraphView const &,
-                     DiGraphPatternMatch const &,
-                     F const &additional_criterion);
-
 bool is_singleton_pattern(OpenMultiDiGraphView const &);
+
+bool assignment_satisfies(ParallelComputationGraph const &,
+                          GraphPattern const &,
+                          DiGraphPatternMatch const &);
 
 } // namespace FlexFlow
