+++ lib/substitutions/include/substitutions/substitution.h
@@ -1,0 +1,18 @@
+#ifndef _FLEXFLOW_SUBSTITUTIONS_SUBSTITUTION_H
+#define _FLEXFLOW_SUBSTITUTIONS_SUBSTITUTION_H
+
+#include "graph_pattern.h"
+#include "output_graph.h"
+
+namespace FlexFlow {
+
+struct Substitution {
+  GraphPattern input_graph;
+  OutputGraph output_graph;
+};
+
+ParallelComputationGraph apply_substitution(ParallelComputationGraph const &, Substitution const &, DiGraphPatternMatch const &);
+
+}
+
+#endif
