+++ lib/substitutions/src/operator_attributes.cc
@@ -1,67 +1,65 @@
 #include "substitutions/get_attribute.h"
-#include "substitutions/substitutions_v2.h"
 
 namespace FlexFlow {
-namespace substitutions {
 
-tl::optional<OperatorAttributeValue> get_attribute(AggregateAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(AggregateAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(AggregateSpecAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(AggregateSpecAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(BatchMatmulAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(BatchMatmulAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(CastAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(CastAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::DATA_TYPE:
       return p.dtype;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(CombineAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(CombineAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::PARALLEL_OP_DIM:
       return p.combine_legion_dim;
     case OperatorAttributeKey::PARALLEL_DIM:
       return p.combine_degree;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(ConcatAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(ConcatAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::AXIS:
       return p.axis;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(Conv2DAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(Conv2DAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::KERNEL_H:
       return p.kernel_h;
@@ -82,38 +80,38 @@ tl::optional<OperatorAttributeValue> get_attribute(Conv2DAttrs const &p,
     case OperatorAttributeKey::USE_BIAS:
       return p.use_bias;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(ElementBinaryAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(ElementBinaryAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(ElementUnaryAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(ElementUnaryAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::SCALAR:
       return p.scalar;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(DropoutAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(DropoutAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(EmbeddingAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(EmbeddingAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::DATA_TYPE:
       return p.data_type;
@@ -124,46 +122,46 @@ tl::optional<OperatorAttributeValue> get_attribute(EmbeddingAttrs const &p,
     case OperatorAttributeKey::OUT_CHANNELS:
       return p.out_channels;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(FlatAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(FlatAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(GatherAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(GatherAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::AXIS:
       return p.legion_dim;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(Group_byAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(Group_byAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(LayerNormAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(LayerNormAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(LinearAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(LinearAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::OUT_CHANNELS:
       return p.out_channels;
@@ -174,24 +172,24 @@ tl::optional<OperatorAttributeValue> get_attribute(LinearAttrs const &p,
     case OperatorAttributeKey::ACTIVATION:
       return p.activation;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue>
-    get_attribute(MultiHeadAttentionAttrs const &p, OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(MultiHeadAttentionAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::NUM_HEADS:
       return p.num_heads;
     case OperatorAttributeKey::USE_BIAS:
       return p.bias;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(Pool2DAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(Pool2DAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::KERNEL_H:
       return p.kernel_h;
@@ -210,105 +208,105 @@ tl::optional<OperatorAttributeValue> get_attribute(Pool2DAttrs const &p,
     case OperatorAttributeKey::ACTIVATION:
       return p.activation;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(ReduceAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(ReduceAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(ReductionAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(ReductionAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::PARALLEL_OP_DIM:
       return p.reduction_legion_dim;
     case OperatorAttributeKey::PARALLEL_OP_DEGREE:
       return p.reduction_degree;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(RepartitionAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(RepartitionAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::PARALLEL_OP_DIM:
       return p.repartition_legion_dim;
     case OperatorAttributeKey::PARALLEL_OP_DEGREE:
       return p.repartition_degree;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(ReplicateAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(ReplicateAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::PARALLEL_OP_DIM:
       return p.replicate_legion_dim;
     case OperatorAttributeKey::PARALLEL_OP_DEGREE:
       return p.replicate_degree;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(ReshapeAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(ReshapeAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(SplitAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(SplitAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::AXIS:
       return p.legion_axis;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(SoftmaxAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(SoftmaxAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::AXIS:
       return p.dim;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(TopKAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(TopKAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue> get_attribute(TransposeAttrs const &p,
-                                                   OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(TransposeAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     case OperatorAttributeKey::PERMUTATION:
       return p.perm;
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
-tl::optional<OperatorAttributeValue>
-    get_attribute(FusedParallelOpAttrs const &p, OperatorAttributeKey key) {
+optional<OperatorAttributeValue> get_attribute(FusedParallelOpAttrs const &p,
+                                               OperatorAttributeKey key) {
   switch (key) {
     default:
-      return tl::nullopt;
+      return nullopt;
   }
 }
 
@@ -316,7 +314,7 @@ struct GetAttribute {
   GetAttribute(OperatorAttributeKey key) : key(key) {}
 
   template <typename T>
-  tl::optional<OperatorAttributeValue> operator()(T const &t) {
+  optional<OperatorAttributeValue> operator()(T const &t) {
     return get_attribute(t, this->key);
   }
 
@@ -324,10 +322,9 @@ private:
   OperatorAttributeKey key;
 };
 
-tl::optional<OperatorAttributeValue> get_attribute(PCGOperatorAttrs const &p,
-                                                   OperatorAttributeKey key) {
-  return mpark::visit(GetAttribute(key), p);
+optional<OperatorAttributeValue> get_attribute(PCGOperatorAttrs const &p,
+                                               OperatorAttributeKey key) {
+  return visit(GetAttribute(key), p);
 }
 
-} // namespace substitutions
 } // namespace FlexFlow
